develop

+++ src/main/kotlin/me/lee/bloger/http/PageResult.kt
@@ -2,6 +2,6 @@ package me.lee.bloger.http
 
 import java.io.Serializable
 
-data class PageResult(val current: Long,
+data class PageResult(val current: Int,
                       val total: Long,
                       val results: Iterable<*>) : Serializable
